dropped the mixlib things in favor of the tty modules;  working with tty-config its still not where I want it but it does allow for envar override of defaults; not sure if that is a one time thing or an on going thing.  tty-option has not been tweaked yet so no cli functions work.  set up for a command structure.
