+++ lib/sqa.rb
@@ -1,3 +1,4 @@
+# lib/sqa.rb
 # frozen_string_literal: true
 
 require 'active_support'
@@ -5,11 +6,8 @@ require 'active_support/core_ext/string'
 require 'daru'
 require 'date'
 require 'descriptive_statistics'
-require 'mixlib/cli'
-require 'mixlib/config'
 require 'nenv'
 require 'pathname'
-require "version_gem"
 
 unless defined?(HOME)
 	HOME = Pathname.new(Nenv.home)
@@ -17,41 +15,8 @@ end
 
 
 module SQA
-	Signal = {
-		hold: 0,
-		buy: 	1,
-		sell: 2
-	}.freeze
-
-	Trend = {
-		up: 	0,
-		down: 1
-	}.freeze
-
-	Swing = {
-		valley: 0,
-		peak: 	1,
-	}.freeze
-
-	module Config
-    extend Mixlib::Config
-    config_strict_mode true
-
-    default :data_dir,  					HOME + "sqa_data"
-    default :plotting_library, 		:gruff  # TODO: use svg-graph
-    default :lazy_update,  				false
-    default :portfolio_filename,	"portfolio.csv"
-    default :trades_filename,    	"trades.csv"
-
-  	default :log_level, 					:info
-  	default :config_filepath, 		"~/.sqa.rb"
-  	default :portfolio_filename, 	"portfolio.csv"
-  	default :trades_filename, 		"trades.csv"
-
-	end
-
 	def self.init
-		SQA::CLI.new.run if defined? SQA::CLI
+		CLI.new.run if defined? CLI
 
 		Config.config_file 		= Pathname.new homify(Config.config_file)
 
@@ -59,8 +24,8 @@ module SQA
 
 		Config.data_dir 			= Pathname.new homify(Config.data_dir)
 
-		Daru.lazy_update 			= Config.lazy_update
-		Daru.plotting_library = Config.plotting_library
+		Daru.lazy_update 			= Config.lazy
+		Daru.plotting_library = Config.plotting_lib
 
 		nil
 	end
@@ -72,6 +37,8 @@ end
 
 # require_relative "patches/daru" # TODO: extract Daru::DataFrame in new gem sqa-data_frame
 
+require_relative "sqa/config"
+require_relative "sqa/constants"
 require_relative "sqa/data_frame"
 require_relative "sqa/errors"
 require_relative "sqa/indicator"
