+++ lib/sqa/command.rb
@@ -1,0 +1,8 @@
+# lib/sqa/command.rb
+
+module SQA
+	module Command
+	end
+end
+
+require_relative 'command/analyze'
